Add files via upload
This is a small project which I have done. In this project, I have used the Django framework of python to make a webpage of an imaginary fruit shop. This is my first time playing around with Django. Previously I have made websites with HTML, CSS, JS, Bootstrap and some automated ones on wix, Here I have worked with the Django Admin Panel. This site offers to manage the items very easily. You just have to add the items from an interactive UI provided by the Django Admin panel, and all the items will be stored on the DBSQL file. You can see the site preview in this project.JPG file.